This Website is now fully responsive.

+++ project/delete_image.php
@@ -1,7 +1,20 @@
 <?php
 session_start();
+if (!$_SESSION['is_login']){
+
+    // Redirect to another page
+    header("Location: login.html");
+    exit();
+    
+    // Flush the output buffer
+    ob_end_flush();
+}
+
+
 $filename = $_GET['delete_file'];
 if (file_exists($filename)) {
+   
+    // delete file
     unlink($filename);
     $_SESSION['msg']  ='File '.$filename.' has been deleted';
     $_SESSION['color'] = "text-success";
